♻️ Refactors shared model code into pydantic layer
Moves shared modules related to DynamoDB and EventBridge into a dedicated 'pydanticde' layer.

This change promotes better code organization and encapsulation by grouping Pydantic-related models and utilities together.
It simplifies dependency management and reduces potential circular dependencies.

This ensures that the functions rely on the layer instead of the _shared folder.
