+++ functions/_layers/pydantic/python/pydanticde/EventBridge.py
