+++ functions/functions/customers/lambda_customers.py
@@ -7,13 +7,12 @@ from datetime import datetime, timezone
 import boto3
 from boto3.dynamodb.conditions import Key
 from botocore.exceptions import ClientError
-from ksuid import KsuidMs
+from ksuid import KsuidMs # utils layer
+from pydanticde.EventBridge import triggerEBEvent # pydantic layer
 
 from _shared.parser import parse_event, validate_event
 from _shared.DecimalEncoder import DecimalEncoder
 from _shared.naming import getOrganisationTableName
-from _shared.EventBridge import triggerEBEvent
-
 
 logger = logging.getLogger()
 logger.setLevel("INFO")
